reasonably working chat but some features still missing
need to put more stuff back also ignore a lof of the chatbox_ files except commands and comman_registration
